WIP: Prototype restconf.common.errors.ErrorPath
We really want to attach EffectiveModelContext to each path, which
highlights where this conversion actually occurs.

The other side of that coin is that this is a purely-server operation,
hence RestconfDocumentedException is really a RestconfServerException,
and it it should be checked and it should not leak outside of a server
implementation (e.g. there should not be an
RestconfDocumentedExceptionMapper, but rather the conversion should
happen explicitly in JAX-RS implementation -- which does have the
required invocation context from the request path.

So this patch will probably be amandoned, but it shows one way of
approaching the YangInstanceIdentifier problem.

There is another down side: we end up instantiating codecs, which we can
already acquire from a DatabindContext and that thing is definitely a
restconf-server thing.

JIRA: NETCONF-1188
Change-Id: I8161f7d2eb41a72e4579e17c771939724c052830
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>