Set accurate YANG Patch status on exception
If a CRUD transaction on the device fails, the specific
failed transaction should be displayed in the PatchStatusEntity.

This patch throws RestconfDocumentedException if a transaction result
contains an RpcError. The RestconfDocumentedException is then linked
with PatchStatusEntity by transactionNumber.
In case of a failed transaction, only the PatchStatusEntity
containing the exception will be displayed.

container edit-status {
           description
             "This container will be present if there are
              edit-specific status responses to report.
              If all edits succeeded and the 'global-status'
              returned is 'ok', then a server MAY omit this
              container.";

JIRA: NETCONF-1197
Change-Id: Idf2871a686a7f5b3e948cdf55da52a34719f7cf2
Signed-off-by: Peter Suna <peter.suna@pantheon.tech>