Remove blocking call from AsyncSshHandler
Remove the ConnectFuture verify() call from the
AsyncSshHandler connect() method.

Implement a method to time out the connection feature
using the Timer class, which creates an asynchronous
thread and does not block other connections.

The SshClient includes its own timeout for connections
with the default value set to 60 seconds.
Cancelling the sshConnectFuture did not halt this process.
To ensure synchronization with our connection timeout
for each device, it is now set to the same duration
as configured in the ConnectTimeoutMillis field.
(Default: 20000ms)

JIRA: NETCONF-1167
Change-Id: Ib9f4211020a165e093bd3a5080e221e70af49ca8
Signed-off-by: Peter Suna <peter.suna@pantheon.tech>