Remove yang.model.api fakes for /restconf/operations
We are not emitting operatiions content through NormalizedNodes and
thefore do not need to mock anything.

For MD-SAL binding this use case is already handled by the parser,
which creates the magic nodes as needed.

For NETCONF binding this can potentially be lacking, as the remote
EffectiveModelContext may not even have ietf-restconf -- it might be
useful there, but we will re-introduce it if need be.

JIRA: NETCONF-773
Change-Id: Ia3f0f07804702d8046a775634a861e92c1761593
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>