Synchronize device timeout with SshClient
The SshClient includes its own timeout for connections
with the default value set to 60 seconds.
Cancelling the sshConnectFuture did not halt this process.
To ensure synchronization with our connection timeout
for each device, it is now set to the same duration
as configured in the ConnectTimeoutMillis field.
(Default: 20000ms)

JIRA: NETCONF-1167
Change-Id: I050ee3c232174d09d4b386782b48ef8f4b2ce7e4
Signed-off-by: Peter Suna <peter.suna@pantheon.tech>