Do not hard-code base path for streams
Base path for streams was hard-coded with assumption that restconf
base path is always configured to be "rests".

There are some users that configures different paths for restconf,
for example lighty.io.

This patch changes this so that we get actual base path from uriInfo
parameter.

See also:
https://github.com/PANTHEONtech/lighty/issues/1682

JIRA: NETCONF-1220
Change-Id: I1c57ff3f4de800a73f364d8d520bb6a82ed2c1ae
Signed-off-by: tobias.pobocik <tobias.pobocik@pantheon.tech>
Signed-off-by: Ivan Hrasko <ivan.hrasko@pantheon.tech>