Fix AsyncSshHandlerTest random test failure
In some rare cases, the creation of a mocked ConnectFuture
may exceed the specified 1000ms timeout set in the
connectTimeoutMillis variable. When this occurs, the test fails
due to a timeout in the AsyncSshHandler class.

To address this issue, the initialization of ConnectFuture
has been moved into the setUp method.

JIRA: NETCONF-1167
Change-Id: Iacbb03f9fc41805cdb43ca43e46cc8fc76747838
Signed-off-by: Peter Suna <peter.suna@pantheon.tech>