WIP: Clean up NetconfMessageTransformer constants
We have a ton of constants relating to naming elements which hark back
to when we did not have Binding representation of RPCs and anyxmls.

These days we have pre-generated constants, so use those, which makes
for looser code coupling and better expressiveness in terms of what we
are doing.

Change-Id: Ied0bda7c3c78cdaa0c59bbca6f1aa9b87342848d
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>