Introduce NetconfTimer
The wiring of a global timer is not quite what we want, as it makes
tracking down where and how it is used quite hard.

Add an explicit NetconfTimer interface and a default implementation to
provide indirection in from on the Netty timer.

This has the benefit of reigning in the lifecycle of the timer, so we
actually end up sharing it across components properly.

JIRA: NETCONF-590
Change-Id: Ib58c9f5668ec859d68432116fe63f954f6eae08e
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>