Split out keystore-legacy
Tracking down the lifecycle of SslContexts brings us to our handling of
key material.

This is currently tangled netconf-client-mdsal using our home-grown (and
problematic) model.

Split out netconf-keystore.yang and the baseline implementation into
keystore-legacy. Now that it sits side-by-side with keystore-api, we can
compare the two.

Since this change requires us to intercept a hidden object, rework the
test in terms of JUnit5.

JIRA: NETCONF-1237
Change-Id: Id9d410e88ec588e148c5f1dff3aad574b3cc8328
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>