Remove callhome-server
The server/provider split does not make sense here -- we end up with
ping-pong through callbacks and whatnot.

Bring the thwo parts into a single component, as callhome-server touches
topology anyway.

While this is not directly tied to the SSL context work, callhome-server
is monitoring netconf-keystore, so the lifecycle needs to tied together
correctly.

JIRA: NETCONF-1243
Change-Id: I75b0b28202a03105169ef675fba7713561eacd68
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>