WIP: Eliminate SchemaResourcesDTO
The process of setting up EffectiveModelContext is rather arcane, going
through multiple indirections, most of which are not really necessary.

This patch introduces DeviceNetconfSchemaProvider to replace the plain
SchemaResourcesDTO, centralizing the process into a single class.

As we now have a stricter lifecycle, we end up using invokeNetconf()
instead of invokeRpc(), making it clear we are only invoking base RPCs.

// FIXME: fix topology-singleton
// FIXME: processingExecutor wiring

JIRA: NETCONF-840
Change-Id: I14d41f9f78332d2da9681de89a931dc01cd0829f
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>