WIP: Refactor NetconfDeviceSchemas
NetconfDeviceSchemas is currently largely useless, as of its two
implementations only one really works.

This patch reworks it to be a record holding information about how the
EffectiveModelContext is to be assembled and moves processing to
NetconfStateSchemasResolverImpl, making it a private detail in process
of doing so.

The checks for yang-library are disabled and will be re-enabled with a
new implementation, which will act as a counter-balance to the current
HELLO+Monitoring implementation.

JIRA: NETCONF-840
Change-Id: Ie0e46656ae022cffbe6724738bd982d51e3d80da
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>