Fixup yanglib-mdsal-writer
yanglib-mdsal-writer does not write out module-state by default, which
our CSIT expects. We also should be friendly to RFC7895 clients and have
this content by default.

Furthermore the @Inject needs to be separated from @Activate, as we do
not have OSGi configuration in javax.inject world. Use a compatible
writeLegacy=true default and Optional for URL provider.

Moreover there is a bug where the initial update would not be written if
the listener is executed before registration is stored into the object
field. Fix that by using an explicit boolean to track the closed state.

Finally enable multiple Urls to be reported in the RFC8525 container.
For RFC7895 compatibility we pick the first Url in iteration order.

JIRA: NETCONF-1277
Change-Id: I8cbb2af5e4e7b8ec713126bad981fb5a38c4240f
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>