Error on using DOMDataTreeWriteTransaction#put
During the integration testing phase of the NetconfDeviceTopologyAdapter,
a specific issue was identified with the DOMDataTreeWriteTransaction#put method.
When executing tests in a predefined order using the @Order annotation,
the put method unexpectedly throws an error. This issue does not occur
with the merge operation, which executes successfully under the same conditions.

JIRA: NETCONF-1278
Change-Id: I56e2468d6172ef5062e2ee4661857dd5ee97703a
Signed-off-by: Yaroslav Lastivka <yaroslav.lastivka@pantheon.tech>