OpenApi for TransportPCE with controller resources fix
Openapi did not fully work for TransportPCE when we selected
"Controller resource - Restconf RFC 8040". This was due to multiple
generation of the API for models with a different revision, now it
is only generated for the latest version.

Added tests for this case which show that the models are loaded
correctly and targeting to a model with an old revision returns the
correct output.

Small refactor of naming in BaseYangOpenApiGenerator - schemaContext
renamed to modelContext.

JIRA: NETCONF-1279
Change-Id: I3042c51715eddfaabe0a1179f1fd0d6cabc9c976
Signed-off-by: lubos-cicut <lubos.cicut@pantheon.tech>