Introduce restconf.api.ConsumableBody
The server-side RequestBody contains the idea it is backed by an
InputStream, making it a mirror of a FormattableBody.

Expose ConsumableBody, which captures that essence, making RequestBody
its subclass. We give up use of VarHandles in favor of an
AtomicReference, which is more friendly to static analysis and human
understanding.

JIRA: NETCONF-773
Change-Id: Icc34a60da460c258d13591df7d2fbf68416a9778
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>