Introduce restconf.server.api.HttpGetResource
HttpGetResource is a useful tool which forces all implementations to
behave in the same manner.

We want httpGet() to return a FormattableBody, which we achieve here for
both OperationsResource and YangLibraryVersionResource.

The latter also gives a useful utility in NormalizedFormattableBody,
which captures the essence of NormalizedNodePayload sans special
handling.

JIRA: NETCONF-773
Change-Id: I48f2c7dec979639c3466024677a36a89919281b7
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>