WIP: Disconnect RestconfServer from RestconfFuture
We are using a RestconfDocumentedException to report errors and map them
to responses -- requiring a databind capture to format in JAX-RS and
further weird tricks.

Make sure the JAX-RS layer does not have to deal with
RestconfDocumentedException by introducing an synchronous interface
which always produces a ServerResponse structure.

The mapping of RestconfDocumentedException a ServerResponse.Failure is
now the responsibility of RestconfServer.

// FIXME: add a utility to create failures

JIRA: NETCONF-1188
Change-Id: I1733374f38c645c25ab6a98ef6590530fbec17d0
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>