WIP: Remove use of RestconfError et al.
RestconfError is a flawed design, for which we have a ServerError
replacement. Deprecate it for removal, as well as all
restconf.common.errors contracts.

This is forcing us to re-think server request completion, which we do by
turning ServerRequest into something that can be completed with either a
successful DTO or an error.

JIRA: NETCONF-1188
Change-Id: I085592594aae5d86f73b79fd4f623008c3907318
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>