Enhance Netconf message handling
This patch is for debugging and testing purposes only.
Enhance the `onMessage` method in the NetconfServerSession handler to
include detailed response handling and error logging for outgoing
Netconf messages. This update introduces a listener to the
`sendMessage` call, which logs failures to send messages and
confirms successful transmissions. The changes aim to improve
debugging capabilities by providing specific message IDs for
each failed or successful send operation, aiding in better
traceability and error handling in the communication process.

NETCONF-1261
Change-Id: I088da6832d3ab267717ee122d53371bf05e8360c
Signed-off-by: Yaroslav Lastivka <yaroslav.lastivka@pantheon.tech>