Openapi: Missing schema for node processed by another parent test
If the same node was a child of more than one parent, only the first
one was processed.

Parent patch fixes this problem. Now when node is processed it will be
saved with the current parent. This patch tests if generated document
is as expected.

Created new test to simplify testing of new reported bugs.

JIRA: NETCONF-1309
Change-Id: I7cc72f5d12286154056da6f5b551a11429518332
Signed-off-by: lubos-cicut <lubos.cicut@pantheon.tech>