Move client config creation to lockedConnect method
NetconfNodeHandler has been refactored to handle client configuration
creation within the NetconfNodeHandler#lockedConnect method.
This approach allows nodeHandler to be successfully created even with
an invalid encrypted password and added into activeConnectors.
If the device has the max-connection-attempts parameter, ODL will attempt
to establish the connection a specified number of times.
If unsuccessful, the status will be changed to unable-to-connect.
On exception, the device's operational data is removed using a DELETE request.

JIRA: NETCONF-1317
Change-Id: I3e13d2c904e01ddf195fc9d9a057c0f1e0828a66
Signed-off-by: Yaroslav Lastivka <yaroslav.lastivka@pantheon.tech>