Refactor ServerRequest
Our invocation currently routes responses via RestconfFuture and then
via RestconfCallback.

This is quite superfluous indirection, which we eliminate by expading
ServerRequest to being an entity that can be completed -- just like a
RestconfCallback.

We then create a specialized JaxRsServerRequest, which is essentially
what JaxRsRestconfCallback used to be, and use it instead of
RestconfFuture in RestconfServer.

JIRA: NETCONF-1188
Change-Id: I870d2dd444f43d71ebb2d249e76097af6d39c33f
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>