Make UncancellableFuture always non cancellable
Clean up UncancellableFuture.java and make uncancellation
to always true.

UncancellableFuture is a future that cannot be cancelled. It is used
to communicate with the device and wait for a response. The purpose of
this, is mainly documentary, because once the message is sent,
the future only wait for a response, which will always come.

JIRA: NETCONF-1336
Change-Id: I7142aa07bb0518f8163b6c56496872b9bd674f58
Signed-off-by: lubos-cicut <lubos.cicut@pantheon.tech>