Fix TCPClient.connect() initialization
The firing of ChannelFutureListeners and handler callbacks is slightly
diifferent in that listeners may be fired in a separate pass, but
handler callbacks are guaranteed to fire during event reception.

We really want to fire when the channel becomes active, similarly to the
no-op we have via ConnectChannelInitializer -- thus immediately updating
the pipeline.

This ensures that NettyIoSession.channelActive() has been invoked before
Netty can deliver any events to it.

JIRA: NETCONF-1348
Change-Id: I68a718d1870a572dc6a1a8bafd5c7b31e6ee4ece
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>