Refactor ServerRpcOperations implementations
We support both local and DOMRpcService-mediated RPC implementations,
but have the relationship rather hard-coded.

InterceptingServerRpcOperations provides the ability to look up a
RpcImplementation and delegate to it. If the lookup fails, the request
will be forwarded to a downstream ServerRpcOperations.

A default downstream is provided as NotSupportedServerRpcOperations,
which always completes the request with an ErrorTag.NOT_SUPPORTED
error.

We also split out DOMRpcService-based implementation into
MdsalServerRpcOperations and have RestconfStrategy instantiate a
combination of these two.

The same is true for a DOMActionService-based implementation, named
MdsalServerActionOperations.

JIRA: NETCONF-1331
Change-Id: Icdf1690a816a67b972eb331e56db038cd0474e1e
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>