WIP: Split out MdsalServerStrategy
MdsalRestconfStrategy is acting as the root of a MdsalRestconfServer,
which is really a historical thing.

MdsalServerStrategy is a new class which takes over the role of the root
ServerStrategy, with a clear relationship with MdsalRestconfServer.

This moves a chunk of logic out of (Mdsal)RestconfStrategy, bringing
RestconfStrategy making it just an implementation of
ServerDataOperations.

JIRA: NETCONF-1331
Change-Id: I7fce7a7f1b857f2c4ce45722831dd2d5404a8be4
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>