Refactor FrameEncoder
We have MessageEncoder and FrameEncoder working in tandem in the
pipeline. While this is fine and dandy from design perspective, it
leaves a bit left to be desired.

The problem is that we end up buffering the raw content of the message
into one ByteBuf (in MessageEncoder) and then copy it to another ByteBuf
(in FrameEncoder).

This is wasteful for FramingMechanism.EOM, which just wants to write the
message out and then append a fre more bytes.

Refactor FrameEncoder into FramingSupport class, which exposes a
writeMessage(), which takes a MessageWriter and a NetconfMessage and
encodes/frames it according to its rules as an atomic operation.

This removes double buffering from EOM framing, but keeps things as they
are for CHUNK framing -- with a note to improve it later.

Change-Id: Ied01af2d6dd106a8d2b98198bde911a5ea09ae08
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>