WIP: Split MessageEncoder up
MessageEncoder has currently two functions:
- it receives a NetconfMessage from the pipeline
- it encodes it into a ByteBuf
- it passes the ByteBuf down the pipeline

Refactor MessageEncoder.encodeTo() to a separate class, MessageWriter,
which handles the details of how to write a NetconfMessage into an
OutputStream.

This makes MessageEncoder a final class, which has the notion of which
MessageWriter it is currently using -- and that can be switched from the
pipeline context.

This completely eliminates the need for modifying the pipeline when
switch XML/EXI and provides for nice integration surface for further
consolidation.

This entire exercise means we have only four objects dealing with
default XML encoding -- be it pretty-printing or otherwise.

Change-Id: I8bb4e75465da26e655226c5c83853adc9c47b670
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>