Refactor delayed MessageWriter replacement
AbstractNetconfSession has a Runnable extension to allow for replacing
the MessageWriter after the next message is processed.

This is really related to DefaultStartExi, which is driving this change
-- but the change can only happen *after* the RPC response is encoded
using current writer.

Make this transition more explicit by moving the mechanics to
MessageEncoder.

This improves message send, as we do not need to hook to every write
just to update it. It also makes it more accurate: the writer needs to
be switched just after the message is *encoded*, not when is written to
wire.

Change-Id: I992e8f0ea11bf5507dbed966cea1fe608ccb6f04
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>