Split out JavaCommonCounters
We are maintaining a native view of CommonCounters, which are defined to
be wrapping unsigned 32bit counters.

Our intermal use does not correctly truncate these, as the original code
got confused by 'Long' representing 'uint32' in Binding long time ago.

There is no need to store these counters in long form -- we can convert
that on demand via a newly-introduced JavaCommonCounters -- which extend
CommonCounters, i.e. the Binding interface generated for the grouping.

Add two implementations, mutable and immutable, and expose the ability
to increment as MutableJavaCommonCounters.

This allows us to to expose the read-only view from NetconfServerSession
without giving users access to directly modify them -- making the class
more convenient.

One side effect of us implementing CommonCounters is that we can use
SessionBuilder(CommonCounters) constructor, which performs an automatic
initialization to current state.

Change-Id: I9287fb1a8d552f236f649c84cfefdf9433c09a5b
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>