Refactor ThreadLocalTransformers
Rather than being a plain holder class with duplicated code for each
constant, rename the class ThreadLocalTransformer so that it extends
ThreadLocal<Transformer>, being abstract and requiring instances to
provide configure() method.

The instances themselves have a nice place in MessageWriter, as that
is where they are used.

Since we have a dedicated class and constants, these replace the boolean
-- instead we store the reference to the ThreadLocalTransformer instance
which produces corresponding transformers.

The equivalent of getters is provided by ThreadLocalTransformer.get(),
i.e. getPrettyTransformer() becomes PRETTY_TRANSFORMER.get() without any
intermediates.

Since we are in the area, let's not access the thread local immediately
-- expose threadLocalTransformer(), which just does ThreadLocal.get(),
so that subclasses defer access to when they actually need it.

Change-Id: I3e641e2f8fd3406052e04e15af9a6a74247d9777
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>